fix egl no context issue after resume from thumbnail
the Program object will be deleted when the Render
init routine called twice on the same Render object.
Fix this issue by deleting Program object beforehand
at the right time.

Another change is move setRenderState call in AppContext
to appropriate event handler to avoid race in the mainLoop.
A good solution is to combine render loop to the android
looper interface.

Signed-off-by: Wayne Wolf <wayne.ron.wu@googlemail>